app/requirements.txt
aioredis==1.3.1
asgiref==3.3.1
async-timeout==2.0.1
attrs==19.3.0
auditlog3==1.0.1
autobahn==20.7.1
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.1
channels==2.4.0
channels-redis==2.4.2
chardet==3.0.4
codecov==2.0.15
colorama==0.4.4
constantly==15.1.0
coverage==5.5
coveralls==1.10.0
cryptography==3.0
daphne==2.5.0
Django==2.2.14
django-admin-interface==0.16.3
django-admin-thumbnails==0.2.5
django-ckeditor==6.0.0
django-clear-cache==0.3
django-colorfield==0.4.1
django-crum==0.7.5
django-dumpdata-one==0.8.3
django-el-pagination==3.2.4
django-flat-responsive==2.0
django-flat-theme==1.1.4
django-ip==1.0.2
django-ipware==2.1.0
django-js-asset==1.2.2
django-pwa==1.0.10
django-six==1.0.4
django-utils-six==2.0
django-widget-tweaks==1.4.8
django-wkhtmltopdf==3.3.0
djangorestframework==3.12.2
docopt==0.6.2
entrypoints==0.3
et-xmlfile==1.0.1
flake8==3.7.9
freezegun==0.3.14
geoip2==3.0.0
h2==3.2.0
hiredis==1.1.0
hpack==3.0.0
html5lib==1.0.1
hyperframe==5.2.0
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
isodate==0.6.0
jdcal==1.4.1
maxminddb==1.5.2
mccabe==0.6.1
mock==3.0.5
msgpack==0.6.2
msgpack-python==0.5.6
mysqlclient==1.4.6
openpyxl==2.6.4
pathlib==1.0.1
Pillow==7.0.0
priority==1.3.0
psycopg2==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.5.0
pycparser==2.20
pyflakes==2.1.1
pygeoip==0.3.2
PyHamcrest==2.0.2
PyMySQL==0.9.3
pyOpenSSL==19.1.0
PyPDF2==1.26.0
python-dateutil==2.8.1
python-memcached==1.59
pytz==2019.3
qrcode==6.1
redis==2.10.6
reportlab==3.5.42
requests==2.23.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.1
Twisted==20.3.0
twisted-iocpsupport==1.0.1
txaio==20.4.1
typing-extensions==3.7.4.3
Unidecode==1.1.1
urllib3==1.25.8
webencodings==0.5.1
xhtml2pdf==0.2.4
zope.interface==5.1.0